Wip users control

+++ src/controllers/offensiveWords.js
@@ -19,13 +19,13 @@ class OffensiveWordsController {
 
     async deleteWord(req, res, next) {
         try {
+            const { role } = req.user;
             const wordId = req.params.wordId;
-            const word = await OffensiveWordsService.deleteWord(wordId);
-            if (word !== null) {
-                res.status(200).send(word);
-                // res.json(word);
+            if (role === "admin") {
+                const newWord = await OffensiveWordsService.deleteWord(wordId);
+                res.status(200).send(newWord);
             } else {
-                res.status(404).send({ message: "Not found" });
+                res.status(401).send({ message: "No tienes permiso para eliminar esta palabra ofensiva" });
             }
         } catch (err) {
             res.status(404).send(err.message);
@@ -36,12 +36,15 @@ class OffensiveWordsController {
 
     async saveWord(req, res, next) {
         try {
+            const { role } = req.user;
             const word = req.body;
-            const newWord = await OffensiveWordsService.saveWord(word);
-            if (typeof newWord.word != 'string' || typeof newWord.level != 'number') {
+            if (typeof word.word != 'string' || typeof word.level != 'number') {
                 res.status(400).send({ message: "La palabra debe tener los campos word y level" });
-            } else {
+            } if (role === "admin") {
+                const newWord = await OffensiveWordsService.saveWord(word);
                 res.status(200).send(newWord);
+            } else {
+                res.status(401).send({ message: "No tienes permiso para crear palabras ofensivas" });
             }
         } catch (err) {
             res.status(500).send(err.message);
@@ -53,12 +56,15 @@ class OffensiveWordsController {
     async updateWord(req, res, next) {
         try {
             const wordId = req.params.wordId;
-            const newWord = req.body;
-            const word = await OffensiveWordsService.updateWord(wordId, newWord);
-            if (typeof newWord.word != 'string' || typeof newWord.level != 'number') {
+            const word = req.body;
+            const { role } = req.user;
+            if (typeof word.word != 'string' || typeof word.level != 'number') {
                 res.status(400).send({ message: "La palabra debe tener los campos word y level" });
+            } if (role === "admin") {
+                const newWord = await OffensiveWordsService.updateWord(wordId, word);
+                res.status(200).send(newWord);
             } else {
-                res.status(200).send(word);
+                res.status(401).send({ message: "No tienes permiso para modificar palabras ofensivas" });
             }
         } catch (err) {
             res.status(404).send(err.message);
